Rename "Logging" integration tests to "AgentLogs"
