+++ tests/Agent/IntegrationTests/IntegrationTests/AgentLogs/ChangeLogFilenameTests.cs
@@ -7,7 +7,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.Logging
+namespace NewRelic.Agent.IntegrationTests.AgentLogs
 {
     [NetFrameworkTest]
     public abstract class ChangeLogFilenameTestsBase : NewRelicIntegrationTest<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
